It seems to work?
Good progress, I am still not sure if it is proper seeking and fleeing behavior. They all seem to follow the same path which is interesting. I might try add some movement to the obstacles which might add some more interesting behavior. Also not the worst performance which is interesting!