Lots of clean up.
Realized I should not be using a global var from the Game Manager for the weights of each of the behaviors. Rather I could make it a member var of the base class and require all inheritors to use it in their constructors. This also Allows for better testing of weights in editor without having to open the code.

Started work on the separation behavior which is making everything freak out, but I just think this is due to the lack of alignment and cohesion behaviors which I will add soon.

Added headers for improved readability in editor as well.

Code is looking rather clean atm.