Alignment
Implementation of alignment, it is very funky at the moment. The Boids are starting and stopping.

I was able to debug and issue where all the Boids were disappearing into the nether because I am now using a closeNeighbourCount that has the potential to be 0. So I return Vector2.zero if it is zero. Dividing by 0 is bad.

Implemented bounds, but I dont think it is working as desired as it will just avoid its x y pos rather than the whole object. That will need to be fixed if I want to use different shapes.

Lots to do but making progress.